Move the NOTIFICATION_AREAS and NOTIFICATION_GROUPS constants out of datastore.
