+++ assets/js/components/notifications/FirstPartyModeSetupBanner.js
@@ -32,10 +32,8 @@ import { __ } from '@wordpress/i18n';
  */
 import { useDispatch, useSelect } from 'googlesitekit-data';
 import { useShowTooltip } from '../AdminMenuTooltip';
-import {
-	CORE_NOTIFICATIONS,
-	NOTIFICATION_AREAS,
-} from '../../googlesitekit/notifications/datastore/constants';
+import { CORE_NOTIFICATIONS } from '../../googlesitekit/notifications/datastore/constants';
+import { NOTIFICATION_AREAS } from '../../googlesitekit/notifications/constants';
 import { CORE_SITE } from '../../googlesitekit/datastore/site/constants';
 import { CORE_USER } from '../../googlesitekit/datastore/user/constants';
 import { DAY_IN_SECONDS } from '../../util';
