+++ assets/js/components/notifications/SubtleNotifications.js
@@ -20,7 +20,7 @@
  * Internal dependencies
  */
 import Notifications from './Notifications';
-import { NOTIFICATION_AREAS } from '../../googlesitekit/notifications/datastore/constants';
+import { NOTIFICATION_AREAS } from '../../googlesitekit/notifications/constants';
 
 export default function SubtleNotifications() {
 	return <Notifications areaSlug={ NOTIFICATION_AREAS.BANNERS_BELOW_NAV } />;
