+++ assets/js/googlesitekit/notifications/register-defaults.js
@@ -33,12 +33,10 @@ import {
 	VIEW_CONTEXT_SETTINGS,
 	VIEW_CONTEXT_SPLASH,
 } from '../constants';
+import { CORE_NOTIFICATIONS } from './datastore/constants';
 import {
-	CORE_NOTIFICATIONS,
-	NOTIFICATION_AREAS,
 	NOTIFICATION_GROUPS,
-} from './datastore/constants';
-import {
+	NOTIFICATION_AREAS,
 	FPM_HEALTH_CHECK_WARNING_NOTIFICATION_ID,
 	FPM_SETUP_CTA_BANNER_NOTIFICATION,
 	PRIORITY,
