+++ assets/js/modules/pagespeed-insights/index.js
@@ -34,7 +34,7 @@ import DashboardPageSpeedWidget from './components/dashboard/DashboardPageSpeedW
 import PageSpeedInsightsIcon from '../../../svg/graphics/pagespeed-insights.svg';
 import { MODULES_PAGESPEED_INSIGHTS } from './datastore/constants';
 import { MODULE_SLUG_PAGESPEED_INSIGHTS } from './constants';
-import { NOTIFICATION_AREAS } from '../../googlesitekit/notifications/datastore/constants';
+import { NOTIFICATION_AREAS } from '../../googlesitekit/notifications/constants';
 import { VIEW_CONTEXT_MAIN_DASHBOARD } from '../../googlesitekit/constants';
 import SetupSuccessNotification from './components/notifications/SetupSuccessNotification';
